Removed old, now incorrect tests.
At launch, the default image size is now specified in the client by
sending a size of "0". This change makes that the user can now specify
zero as an image size; this value will behave now as requesting to use
the default size and not as specifying a disk size of zero. The tests
testing that the specified disk size is not zero were removed.